some additions
