+++ ydb/core/kqp/opt/physical/kqp_opt_phy_rules.h
@@ -37,6 +37,15 @@ NYql::NNodes::TExprBase KqpBuildStreamIdxLookupJoinStages(NYql::NNodes::TExprBas
 NYql::NNodes::TExprBase KqpRemoveRedundantSortByPk(NYql::NNodes::TExprBase node, NYql::TExprContext& ctx,
     const TKqpOptimizeContext& kqpCtx);
 
+NYql::NNodes::TExprBase KqpBuildTopStageRemoveSort(
+    NYql::NNodes::TExprBase node, 
+    NYql::TExprContext& ctx, 
+    NYql::IOptimizationContext& optCtx, 
+    NYql::TTypeAnnotationContext& typeCtx,
+    const NYql::TParentsMap& parentsMap, 
+    bool allowStageMultiUsage
+);
+
 NYql::NNodes::TExprBase KqpApplyLimitToReadTable(NYql::NNodes::TExprBase node, NYql::TExprContext& ctx,
     const TKqpOptimizeContext& kqpCtx);
