+++ ydb/core/kqp/opt/physical/kqp_opt_phy_sort.cpp
@@ -5,11 +5,13 @@
 #include <ydb/core/kqp/opt/kqp_opt_impl.h>
 
 #include <yql/essentials/core/yql_opt_utils.h>
+#include <ydb/library/yql/dq/type_ann/dq_type_ann.h>
 
 namespace NKikimr::NKqp::NOpt {
 
 using namespace NYql;
 using namespace NYql::NNodes;
+using namespace NYql::NDq;
 
 // Temporary solution, should be replaced with constraints
 // copy-past from old engine algo: https://a.yandex-team.ru/arc_vcs/yql/providers/kikimr/yql_kikimr_opt.cpp?rev=e592a5a9509952f1c29f1ec02343dd4c05fe426d#L122
@@ -100,6 +102,144 @@ TExprBase KqpRemoveRedundantSortByPk(TExprBase node, TExprContext& ctx, const TK
     } else {
         return input;
     }
+}
+
+using namespace NYql::NDq;
+
+bool CompatibleSort(TOptimizerStatistics::TSortColumns& existingOrder, const TCoLambda& keySelector, const TExprBase& sortDirections, TVector<TString>& sortKeys) {
+    if (auto body = keySelector.Body().Maybe<TCoMember>()) {
+        auto attrRef = body.Cast().Name().StringValue();
+        auto attrName = existingOrder.Columns[0];
+        auto attrNameWithAlias = existingOrder.Aliases[0] + "." + attrName;
+        if (attrName == attrRef || attrNameWithAlias == attrRef){
+            auto sortValue = sortDirections.Cast<TCoDataCtor>().Literal().Value();
+            if (FromString<bool>(sortValue)) {
+                sortKeys.push_back(attrRef);
+                return true;
+            }
+        }
+    }
+    else if (auto body = keySelector.Body().Maybe<TExprList>()) {
+        if (body.Cast().Size() > existingOrder.Columns.size()) {
+            return false;
+        }
+
+        bool allMatched = false;
+        auto dirs = sortDirections.Cast<TExprList>();
+        for (size_t i=0; i < body.Cast().Size(); i++) {
+            allMatched = false;
+            auto item = body.Cast().Item(i);
+            if (auto member = item.Maybe<TCoMember>()) {
+                auto attrRef = member.Cast().Name().StringValue();
+                auto attrName = existingOrder.Columns[i];
+                auto attrNameWithAlias = existingOrder.Aliases[i] + "." + attrName;
+                if (attrName == attrRef || attrNameWithAlias == attrRef){
+                    auto sortValue = dirs.Item(i).Cast<TCoDataCtor>().Literal().Value();
+                    if (FromString<bool>(sortValue)) {
+                        sortKeys.push_back(attrRef);
+                        allMatched = true;
+                    }
+                }
+            }
+            if (!allMatched) {
+                return false;
+            }
+        }
+        return true;
+    }
+    return false;
+}
+
+TExprBase KqpBuildTopStageRemoveSort(
+    TExprBase node, 
+    TExprContext& ctx, 
+    IOptimizationContext& optCtx, 
+    TTypeAnnotationContext& typeCtx,
+    const TParentsMap& parentsMap, 
+    bool allowStageMultiUsage
+) {
+    if (!node.Maybe<TCoTopBase>().Input().Maybe<TDqCnUnionAll>()) {
+        return node;
+    }
+
+    const auto top = node.Cast<TCoTopBase>();
+    const auto dqUnion = top.Input().Cast<TDqCnUnionAll>();
+
+    if (
+        dqUnion.Output().Stage().Program().Body().Maybe<TCoFlatMap>()
+    ) {
+        auto flatmap = dqUnion.Output().Stage().Program().Body().Cast<TCoFlatMap>();
+        if (IsPassthroughFlatMap(flatmap, nullptr)) {
+            return node;
+            auto input = flatmap.Input();
+            bool isReadTable = input.Maybe<TKqpReadTable>().IsValid();
+            bool isReadTableRanges = input.Maybe<TKqpReadTableRanges>().IsValid() || input.Maybe<TKqpReadOlapTableRanges>().IsValid() ;
+            if (!isReadTable && !isReadTableRanges) {
+                return node;
+            }
+        }
+    }
+
+    auto inputStats = typeCtx.GetStats(dqUnion.Output().Raw());
+    
+    if (!inputStats || !inputStats->SortColumns) {
+        return node;
+    }
+
+    if (!IsSingleConsumerConnection(dqUnion, parentsMap, allowStageMultiUsage)) {
+        return node;
+    }
+
+    if (!CanPushDqExpr(top.Count(), dqUnion) || !CanPushDqExpr(top.KeySelectorLambda(), dqUnion)) {
+        return node;
+    }
+
+    if (auto connToPushableStage = DqBuildPushableStage(dqUnion, ctx)) {
+        return TExprBase(ctx.ChangeChild(*node.Raw(), TCoTop::idx_Input, std::move(connToPushableStage)));
+    }
+
+    const auto sortKeySelector = top.KeySelectorLambda();
+    const auto sortDirections = top.SortDirections();
+    TVector<TString> sortKeys;
+
+    if (!CompatibleSort(*inputStats->SortColumns, sortKeySelector, sortDirections, sortKeys)) {
+        return node;
+    }
+
+    auto builder = Build<TDqSortColumnList>(ctx, node.Pos());
+    for (auto columnName : sortKeys ) {
+        builder.Add<TDqSortColumn>()
+            .Column<TCoAtom>().Build(columnName)
+            .SortDirection().Build(TTopSortSettings::AscendingSort)
+            .Build();
+    }
+    auto columnList = builder.Build().Value();
+
+    return Build<TDqCnUnionAll>(ctx, node.Pos())
+        .Output()
+            .Stage<TDqStage>()
+                .Inputs()
+                    .Add<TDqCnMerge>()
+                        .Output()
+                            .Stage(dqUnion.Output().Stage())
+                            .Index(dqUnion.Output().Index())
+                            .Build()
+                        .SortColumns(columnList)
+                        .Build()
+                    .Build()
+                .Program()
+                    .Args({"stream"})
+                    .Body<TCoTake>()
+                        .Input("stream")
+                        .Count(top.Count())
+                        .Build()
+                    .Build()
+                .Settings(NDq::TDqStageSettings::New().BuildNode(ctx, top.Pos()))
+                .Build()
+            .Index().Build(0U)
+            .Build()
+        //.SortColumns(columnList)
+        .Done();
 }
 
 } // namespace NKikimr::NKqp::NOpt
