+++ ydb/library/yql/dq/opt/dq_opt_phy.cpp
@@ -1577,122 +1577,6 @@ TExprBase GetSortDirection(const TExprBase& sortDirections, size_t index) {
 }
 } // End of anonymous namespace
 
-bool CompatibleSort(TOptimizerStatistics::TSortColumns& existingOrder, const TCoLambda& keySelector, const TExprBase& sortDirections, TVector<TString>& sortKeys) {
-    if (auto body = keySelector.Body().Maybe<TCoMember>()) {
-        auto attrRef = body.Cast().Name().StringValue();
-        auto attrName = existingOrder.Columns[0];
-        auto attrNameWithAlias = existingOrder.Aliases[0] + "." + attrName;
-        if (attrName == attrRef || attrNameWithAlias == attrRef){
-            auto sortValue = sortDirections.Cast<TCoDataCtor>().Literal().Value();
-            if (FromString<bool>(sortValue)) {
-                sortKeys.push_back(attrRef);
-                return true;
-            }
-        }
-    }
-    else if (auto body = keySelector.Body().Maybe<TExprList>()) {
-        if (body.Cast().Size() > existingOrder.Columns.size()) {
-            return false;
-        }
-
-        bool allMatched = false;
-        auto dirs = sortDirections.Cast<TExprList>();
-        for (size_t i=0; i < body.Cast().Size(); i++) {
-            allMatched = false;
-            auto item = body.Cast().Item(i);
-            if (auto member = item.Maybe<TCoMember>()) {
-                auto attrRef = member.Cast().Name().StringValue();
-                auto attrName = existingOrder.Columns[i];
-                auto attrNameWithAlias = existingOrder.Aliases[i] + "." + attrName;
-                if (attrName == attrRef || attrNameWithAlias == attrRef){
-                    auto sortValue = dirs.Item(i).Cast<TCoDataCtor>().Literal().Value();
-                    if (FromString<bool>(sortValue)) {
-                        sortKeys.push_back(attrRef);
-                        allMatched = true;
-                    }
-                }
-            }
-            if (!allMatched) {
-                return false;
-            }
-        }
-        return true;
-    }
-    return false;
-}
-
-TExprBase DqBuildTopStageRemoveSort(TExprBase node, TExprContext& ctx, IOptimizationContext& optCtx, TTypeAnnotationContext& typeCtx,
-    const TParentsMap& parentsMap, bool allowStageMultiUsage)
-{
-    if (!node.Maybe<TCoTopBase>().Input().Maybe<TDqCnUnionAll>()) {
-        return node;
-    }
-
-    const auto top = node.Cast<TCoTopBase>();
-    const auto dqUnion = top.Input().Cast<TDqCnUnionAll>();
-
-    auto inputStats = typeCtx.GetStats(dqUnion.Output().Raw());
-    
-    if (!inputStats || !inputStats->SortColumns) {
-        return node;
-    }
-
-    if (!IsSingleConsumerConnection(dqUnion, parentsMap, allowStageMultiUsage)) {
-        return node;
-    }
-
-    if (!CanPushDqExpr(top.Count(), dqUnion) || !CanPushDqExpr(top.KeySelectorLambda(), dqUnion)) {
-        return node;
-    }
-
-    if (auto connToPushableStage = DqBuildPushableStage(dqUnion, ctx)) {
-        return TExprBase(ctx.ChangeChild(*node.Raw(), TCoTop::idx_Input, std::move(connToPushableStage)));
-    }
-
-    const auto sortKeySelector = top.KeySelectorLambda();
-    const auto sortDirections = top.SortDirections();
-    TVector<TString> sortKeys;
-
-    if (!CompatibleSort(*inputStats->SortColumns, sortKeySelector, sortDirections, sortKeys)) {
-        return node;
-    }
-
-    auto builder = Build<TDqSortColumnList>(ctx, node.Pos());
-    for (auto columnName : sortKeys ) {
-        builder.Add<TDqSortColumn>()
-            .Column<TCoAtom>().Build(columnName)
-            .SortDirection().Build(TTopSortSettings::AscendingSort)
-            .Build();
-    }
-    auto columnList = builder.Build().Value();
-
-    return Build<TDqCnUnionAll>(ctx, node.Pos())
-        .Output()
-            .Stage<TDqStage>()
-                .Inputs()
-                    .Add<TDqCnMerge>()
-                        .Output()
-                            .Stage(dqUnion.Output().Stage())
-                            .Index(dqUnion.Output().Index())
-                            .Build()
-                        .SortColumns(columnList)
-                        .Build()
-                    .Build()
-                .Program()
-                    .Args({"stream"})
-                    .Body<TCoTake>()
-                        .Input("stream")
-                        .Count(top.Count())
-                        .Build()
-                    .Build()
-                .Settings(TDqStageSettings().BuildNode(ctx, top.Pos()))
-                .Build()
-            .Index().Build(0U)
-            .Build()
-        //.SortColumns(columnList)
-        .Done();
-}
-
 TExprBase DqBuildTopStage(TExprBase node, TExprContext& ctx, IOptimizationContext& optCtx,
     const TParentsMap& parentsMap, bool allowStageMultiUsage)
 {
